output time in microseconds instead of seconds

+++ src/util.h
@@ -10,7 +10,7 @@ int is_not_option(char* s);
 /* converts a string to a file descriptor int */
 int get_file_desc(char *str);
 
-/* returns the time elapsed */
-double get_time_diff(struct timeval prev, struct timeval curr);
+/* returns the time elapsed in microseconds */
+int get_time_diff(struct timeval prev, struct timeval curr);
 
 #endif /* UTIL_H */
